build fixes

+++ main.cpp
@@ -112,7 +112,7 @@ int main(int argc, char **argv) {
 	call_load_functions();
 
 	//------------ setting up random seed --------------
-	srand(std::time(nullptr));
+	srand((int)std::time(nullptr));
 
 	//------------ create game mode + make current --------------
 	Mode::set_current(std::make_shared< PlayMode >());
